requirements.txt
# GUI library
tkinter

# Database library
mysql-connector-python

# Additional libraries
datetime